+++ src/page-objects/components/exui-case-list.component.ts
@@ -2,8 +2,8 @@ import { Page } from "@playwright/test";
 import { ExuiSpinnerComponent } from "./exui-spinner.component.js";
 
 export class ExuiCaseListComponent {
-  constructor(public page: Page) {}
-  
+  constructor(public readonly page: Page) {}
+
   readonly caseList = this.page.locator("exui-case-list");
   readonly caseListTable = this.page.locator("#search-result table");
   readonly filters = {
@@ -11,10 +11,10 @@ export class ExuiCaseListComponent {
     caseNumberFilter: this.page.locator("#\\[CASE_REFERENCE\\]"),
     caseStateFilter: this.page.locator("select#wb-case-state"),
     applyFilterBtn: this.page.getByTitle("Apply filter"),
-  };
+  } as const;
   readonly resultLinks = this.page.locator("ccd-search-result .govuk-link");
-  private spinnerComponent = new ExuiSpinnerComponent(this.page);
-  
+  private readonly spinnerComponent = new ExuiSpinnerComponent(this.page);
+
   public async searchByCaseName(caseName: string): Promise<void> {
     await this.filters.caseNameFilter.fill(caseName);
     await this.filters.applyFilterBtn.click();
@@ -27,13 +27,13 @@ export class ExuiCaseListComponent {
     await this.spinnerComponent.wait();
   }
 
-  public async searchByCaseState(state: string) {
+  public async searchByCaseState(state: string): Promise<void> {
     await this.filters.caseStateFilter.selectOption(state);
     await this.filters.applyFilterBtn.click();
     await this.spinnerComponent.wait();
   }
 
-  public async selectCaseByIndex(index: number) {
+  public async selectCaseByIndex(index: number): Promise<void> {
     await this.resultLinks.nth(index).click();
     await this.spinnerComponent.wait();
   }