+++ src/utils/browser.utils.ts
@@ -1,4 +1,9 @@
-import { Browser, BrowserContext } from "@playwright/test";
+import {
+  Browser,
+  BrowserContext,
+  BrowserContextOptions,
+  Page,
+} from "@playwright/test";
 
 export class BrowserUtils {
   constructor(protected readonly browser: Browser) {}
@@ -9,11 +14,14 @@ export class BrowserUtils {
    * @param sessionFile {@link string} - optionally provide a session file to use in the new browser context
    *
    */
-  public async openNewBrowserContext(sessionFile?: string) {
-    const browser: Browser = await this.browser.browserType().launch();
-    const context: BrowserContext = await browser.newContext({
-      storageState: sessionFile ? sessionFile : undefined,
-    });
+  public async openNewBrowserContext(sessionFile?: string): Promise<Page> {
+    const contextOptions: BrowserContextOptions = {};
+    if (sessionFile) {
+      contextOptions.storageState = sessionFile;
+    }
+    const context: BrowserContext = await this.browser.newContext(
+      contextOptions
+    );
     return context.newPage();
   }
 }