+++ src/utils/locale.utils.ts
@@ -12,17 +12,12 @@ export class LocaleUtils {
    * Takes a locale and appends it to the current URL, or if a locale is already present it is replaced.
    *
    * @param locale {@link Locale} - enum representing supported locales
-   *
+  *
    */
   public appendLocaleToUrl(locale: Locale): string {
-    const url = this.page.url();
-    const lngRegex = new RegExp("\\?lng=");
-
-    const match = url.match(lngRegex);
-    if (match) {
-      return this.page.url().replace(/\?lng=\w+/, `?lng=${locale}`);
-    }
-    return this.page.url() + `?lng=${locale}`;
+    const currentUrl = new URL(this.page.url());
+    currentUrl.searchParams.set("lng", locale);
+    return currentUrl.toString();
   }
 
   public async navigateWithLocale(locale: Locale): Promise<void> {